test: separate async api contract integration coverage
