+++ docs/architecture/system-overview.md
@@ -112,6 +112,9 @@ contract coverage can evolve independently of the broader platform health suite.
 Evaluation API behavior is also verified through its own dedicated integration module so
 evaluation contract coverage can evolve independently of the broader platform health suite.
 
+Async API behavior is also verified through its own dedicated integration module so async
+contract coverage can evolve independently of the broader platform health suite.
+
 ### Async Runtime
 
 - `src/app/services/async_runtime_status.py`
